refactor: move layout to /lib
