+++ src/cli/commands/create/app.ts
@@ -2,7 +2,7 @@ import { stripIndent } from 'common-tags'
 import * as fs from 'fs-jetpack'
 import prompts from 'prompts'
 import { PackageJson } from 'type-fest'
-import * as Layout from '../../../framework/layout'
+import * as Layout from '../../../lib/layout'
 import { Command } from '../../../lib/cli'
 import { ownPackage } from '../../../lib/own-package'
 import * as Plugin from '../../../lib/plugin'
