+++ src/cli/main.ts
@@ -6,7 +6,7 @@ import { stripIndent } from 'common-tags'
 import * as dotenv from 'dotenv'
 import * as Path from 'path'
 import { isError } from 'util'
-import * as Layout from '../framework/layout'
+import * as Layout from '../lib/layout'
 import { CLI, HelpError } from '../lib/cli'
 import * as ExitSystem from '../lib/exit-system'
 import { fatal, isProcessFromProjectBin } from '../utils'
