+++ src/framework/app.ts
@@ -2,7 +2,7 @@ import * as HTTP from 'http'
 import * as Lo from 'lodash'
 import * as Logger from '../lib/logger'
 import * as Plugin from '../lib/plugin'
-import * as Layout from './layout'
+import * as Layout from '../lib/layout'
 import * as Schema from './schema'
 import * as Server from './server'
 import * as singletonChecks from './singleton-checks'
