+++ src/framework/testing.ts
@@ -3,7 +3,7 @@ import * as Lo from 'lodash'
 import { GraphQLClient } from '../lib/graphql-client'
 import * as Plugin from '../lib/plugin'
 import * as app from './index'
-import * as Layout from './layout'
+import * as Layout from '../lib/layout'
 import * as singletonChecks from './singleton-checks'
 
 type AppClient = {
