+++ src/utils/tsc.ts
@@ -2,7 +2,7 @@ import { stripIndent } from 'common-tags'
 import * as fs from 'fs-jetpack'
 import * as path from 'path'
 import * as ts from 'typescript'
-import { Layout } from '../framework/layout'
+import { Layout } from '../lib/layout'
 import { rootLogger } from './logger'
 
 const log = rootLogger.child('compiler')
