+++ src/watcher/watcher.ts
@@ -1,5 +1,5 @@
 import anymatch from 'anymatch'
-import { saveDataForChildProcess } from '../framework/layout'
+import { saveDataForChildProcess } from '../lib/layout'
 import { clearConsole } from '../utils'
 import { rootLogger } from '../utils/logger'
 import { watch } from './chokidar'
