+++ test/e2e/index.test.ts
@@ -1,5 +1,5 @@
 import { introspectionQuery } from 'graphql'
-import { CONVENTIONAL_SCHEMA_FILE_NAME } from '../../src/framework/layout/schema-modules'
+import { CONVENTIONAL_SCHEMA_FILE_NAME } from '../../src/lib/layout/schema-modules'
 import { setupE2EContext } from '../../src/utils/e2e-testing'
 
 const ctx = setupE2EContext()
